Integrate managed ViewModels and converters
Source-Checkpoint: 97e2822fd8c81071b2d4387ee9547d8defbd0257
Original-Commits: 43
Source attribution is retained in the source files and backup records.

+++ src/CalcManagerManaged/CalcManagerPort/CalculatorHistory.cpp.cs
@@ -68,11 +68,11 @@ public partial class CalculatorHistory
         return (uint)(m_historyItems.Count - 1);
     }
 
-    public bool RemoveItem(uint uIdx)
+    public bool RemoveItem(int uIdx)
     {
         if (uIdx < m_historyItems.Count)
         {
-            m_historyItems.RemoveAt((int)uIdx);
+            m_historyItems.RemoveAt(uIdx);
             //m_historyItems.erase(m_historyItems.begin() + uIdx);
             return true;
         }
